% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_lambda.R
\name{find_lambda}
\alias{find_lambda}
\title{Return the lambda which causes all the points become fused into one cluster.}
\usage{
find_lambda(X)
}
\arguments{
\item{X}{data matrix of n * p}
}
\value{
the biggest lambda
}
\description{
Return the lambda which causes all the points become fused into one cluster.
}
\examples{
X = matrix(rnorm(3*2), 3, 2)
find_lambda(X)
}
